Upgrade Astro from 5 to 6.4.2
