+++ package.json
@@ -1,76 +1,76 @@
 {
-  "name": "fuwari",
-  "type": "module",
-  "version": "0.0.1",
-  "scripts": {
-    "dev": "astro dev",
-    "start": "astro dev",
-    "check": "astro check",
-    "build": "astro build && pagefind --site dist",
-    "preview": "astro preview",
-    "astro": "astro",
-    "type-check": "tsc --noEmit --isolatedDeclarations",
-    "new-post": "node scripts/new-post.js",
-    "format": "biome format --write ./src",
-    "lint": "biome check --write ./src",
-    "preinstall": "npx only-allow pnpm"
-  },
-  "dependencies": {
-    "@astrojs/check": "^0.9.6",
-    "@astrojs/rss": "^4.0.14",
-    "@astrojs/sitemap": "^3.6.0",
-    "@astrojs/svelte": "7.2.3",
-    "@astrojs/tailwind": "^6.0.2",
-    "@expressive-code/core": "^0.41.4",
-    "@expressive-code/plugin-collapsible-sections": "^0.41.4",
-    "@expressive-code/plugin-line-numbers": "^0.41.4",
-    "@fontsource-variable/jetbrains-mono": "^5.2.8",
-    "@fontsource/roboto": "^5.2.9",
-    "@iconify-json/fa6-brands": "^1.2.6",
-    "@iconify-json/fa6-regular": "^1.2.4",
-    "@iconify-json/fa6-solid": "^1.2.4",
-    "@iconify-json/material-symbols": "^1.2.50",
-    "@iconify/svelte": "^4.2.0",
-    "@swup/astro": "^1.7.0",
-    "@tailwindcss/typography": "^0.5.19",
-    "astro": "5.13.10",
-    "astro-expressive-code": "^0.41.4",
-    "astro-icon": "^1.1.5",
-    "hastscript": "^9.0.1",
-    "katex": "^0.16.27",
-    "markdown-it": "^14.1.0",
-    "mdast-util-to-string": "^4.0.0",
-    "overlayscrollbars": "^2.12.0",
-    "pagefind": "^1.4.0",
-    "photoswipe": "^5.4.4",
-    "reading-time": "^1.5.0",
-    "rehype-autolink-headings": "^7.1.0",
-    "rehype-components": "^0.3.0",
-    "rehype-katex": "^7.0.1",
-    "rehype-slug": "^6.0.0",
-    "remark-directive": "^3.0.1",
-    "remark-directive-rehype": "^0.4.2",
-    "remark-github-admonitions-to-directives": "^1.0.5",
-    "remark-math": "^6.0.0",
-    "remark-sectionize": "^2.1.0",
-    "sanitize-html": "^2.17.0",
-    "sharp": "^0.34.5",
-    "stylus": "^0.64.0",
-    "svelte": "^5.39.8",
-    "tailwindcss": "^3.4.19",
-    "typescript": "^5.9.3",
-    "unist-util-visit": "^5.0.0"
-  },
-  "devDependencies": {
-    "@astrojs/ts-plugin": "^1.10.6",
-    "@biomejs/biome": "2.2.5",
-    "@rollup/plugin-yaml": "^4.1.2",
-    "@types/hast": "^3.0.4",
-    "@types/markdown-it": "^14.1.2",
-    "@types/mdast": "^4.0.4",
-    "@types/sanitize-html": "^2.16.0",
-    "postcss-import": "^16.1.1",
-    "postcss-nesting": "^13.0.2"
-  },
-  "packageManager": "pnpm@9.14.4"
+    "name": "fuwari",
+    "type": "module",
+    "version": "0.0.1",
+    "scripts": {
+        "dev": "astro dev",
+        "start": "astro dev",
+        "check": "astro check",
+        "build": "astro build && pagefind --site dist",
+        "preview": "astro preview",
+        "astro": "astro",
+        "type-check": "tsc --noEmit --isolatedDeclarations",
+        "new-post": "node scripts/new-post.js",
+        "format": "biome format --write ./src",
+        "lint": "biome check --write ./src",
+        "preinstall": "npx only-allow pnpm"
+    },
+    "dependencies": {
+        "@astrojs/check": "^0.9.9",
+        "@astrojs/rss": "^4.0.18",
+        "@astrojs/sitemap": "^3.7.3",
+        "@astrojs/svelte": "8.1.2",
+        "@astrojs/tailwind": "^6.0.2",
+        "@expressive-code/core": "^0.41.4",
+        "@expressive-code/plugin-collapsible-sections": "^0.41.4",
+        "@expressive-code/plugin-line-numbers": "^0.41.4",
+        "@fontsource-variable/jetbrains-mono": "^5.2.8",
+        "@fontsource/roboto": "^5.2.9",
+        "@iconify-json/fa6-brands": "^1.2.6",
+        "@iconify-json/fa6-regular": "^1.2.4",
+        "@iconify-json/fa6-solid": "^1.2.4",
+        "@iconify-json/material-symbols": "^1.2.50",
+        "@iconify/svelte": "^4.2.0",
+        "@swup/astro": "^1.7.0",
+        "@tailwindcss/typography": "^0.5.19",
+        "astro": "6.4.2",
+        "astro-expressive-code": "^0.41.4",
+        "astro-icon": "^1.1.5",
+        "hastscript": "^9.0.1",
+        "katex": "^0.16.27",
+        "markdown-it": "^14.1.0",
+        "mdast-util-to-string": "^4.0.0",
+        "overlayscrollbars": "^2.12.0",
+        "pagefind": "^1.4.0",
+        "photoswipe": "^5.4.4",
+        "reading-time": "^1.5.0",
+        "rehype-autolink-headings": "^7.1.0",
+        "rehype-components": "^0.3.0",
+        "rehype-katex": "^7.0.1",
+        "rehype-slug": "^6.0.0",
+        "remark-directive": "^3.0.1",
+        "remark-directive-rehype": "^0.4.2",
+        "remark-github-admonitions-to-directives": "^1.0.5",
+        "remark-math": "^6.0.0",
+        "remark-sectionize": "^2.1.0",
+        "sanitize-html": "^2.17.0",
+        "sharp": "^0.34.5",
+        "stylus": "^0.64.0",
+        "svelte": "^5.39.8",
+        "tailwindcss": "^3.4.19",
+        "typescript": "^5.9.3",
+        "unist-util-visit": "^5.0.0"
+    },
+    "devDependencies": {
+        "@astrojs/ts-plugin": "^1.10.9",
+        "@biomejs/biome": "2.2.5",
+        "@rollup/plugin-yaml": "^4.1.2",
+        "@types/hast": "^3.0.4",
+        "@types/markdown-it": "^14.1.2",
+        "@types/mdast": "^4.0.4",
+        "@types/sanitize-html": "^2.16.0",
+        "postcss-import": "^16.1.1",
+        "postcss-nesting": "^13.0.2"
+    },
+    "packageManager": "pnpm@9.14.4"
 }
