+++ src/components/misc/ImageWrapper.astro
@@ -33,7 +33,9 @@ if (isLocal) {
 		import: "default",
 	});
 	let normalizedPath = path
-		.normalize(path.join("../../", basePath, src))
+		.normalize(
+			path.join("../../", basePath.replace(/^src\/content\//, "content/"), src),
+		)
 		.replace(/\\/g, "/");
 	const file = files[normalizedPath];
 	if (!file) {
